Default the menubar and toolbar to off in GUI mode
When I was first learning Vim it was helpful to have them on and was one
of the main reasons I used GVim when I did---as a learning tool for
terminal Vim. Now it is one of the main reasons I don't use
GVim---because they take up screen real estate. That's too bad; might as
well default them to off. It's not as if they are hard to bring back if
necessary!